+++ src/reducers/changeState.js
@@ -62,16 +62,7 @@ const databaseReducer = (state, action, type) => {
 	if (action.type === `custom${upperFirst(type)}_Changed`) {
 		let obj = clone(data[type]);
 		if (action.payload) obj = merge(obj, action.payload);
-		if (action.setting && action.setting.length > 0 && !action.setting.includes('All') && type !== 'settings') {
-			let filter = {};
-			Object.keys(obj).forEach(key => {
-				if (obj[key].setting) {
-					if (obj[key].setting.includes('All') || action.setting.some(setting => obj[key].setting.includes(setting))) filter[key] = clone(obj[key]);
-				} else filter[key] = clone(obj[key]);
-			});
-			return filter;
-		}
-		else return obj;
+		return obj;
 	}
 	return state;
 };